Move some functions to seperate modules
